ootb: make use of `default` as serviceaccount name
throughout the whole installation (and CNR notes about namespace setup),
`default` is the name of the serviceaccount, so, this commit moves us
towards sticking to that.

ps.: the `tap` package (used in the profile-based installation) makes
use of no extra configuration for `ootb` with regards to
service_account, which means "default" is used.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>