ootb/scanning: improve scantemplate recomendations
although it's totally possible for one to modify the templates for the
supplychains (or even the supplychains themselves), we currently provide
no guides for doing so.

as such, changing the recommendation here to not explicitly point at
custom creation of scantemplates but rather point at the page where
there's a good description of them.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>