docs(tap-gui): don't say "external" ClusterIssuer
The idea of an "external" (Cluster)Issuer is a specific concept of
cert-manager. It refers to (Cluster)Issuer which are maintained
out-of-tree. This makes it possible to extend cert-manager with
implementations for almost any CA. (See:
https://cert-manager.io/docs/configuration/external/)

Hence, the phrase "external ClusterIssuer" is misleading as it is used
by TAP GUI's TLS docs. Just saying "ClusterIssuer" is better and just as
precise.

Signed-off-by: Max Brauer <mbrauer@vmware.com>