Modifies the backstage versions transform to remove slashes
Slashes apparently can't be used in the template_variables. This updated
script is able to produce the same variables that are currently in the
template_variables file. I re-ran the script and got no diff in that
file when compared to the hand modified variables that had the slashes
removed.

Signed-off-by: Michael Stergianis <mstergianis@vmware.com>